Add VPN monitoring alerts to 2-networking-a/b
The Fast stage 2-networking-* currently adds monitoring dashboards
for VPN metrics. This change proposes also to add the following
monitoring alerts:

* VPN tunnel established
*
[VPN bandwidth](https://cloud.google.com/network-connectivity/docs/vpn/how-to/viewing-logs-metrics#define-bandwidth-alerts)

To configure the alerts, there is a new `alert_config` variable with
defined default values. By default, the alerts are not created. In order
to create the alerts with default values, add the following to
`terraform.tfvars`:

```
alert_config = {
  vpn_tunnel_established = {}
  vpn_tunnel_bandwidth = {}
}
```